DRAFT: heat timeout support
Heat stack in CI job failed due to some Nova issue.

But then apparently yardstick kept running and took 180mins to timeout

https://build.opnfv.org/ci/view/bottlenecks/job/bottlenecks-compass-posca_stress_ping-baremetal-daily-master/16/console

We can add a Heat create timeout and fail faster if there is an error.

The question is how long should we wait for a Heat stack to deploy.  We
can set a default and allow override in the heat context config, if
users make complicated stacks

Change-Id: Ib4c24c852ef2a97396b1567cbea95b8f1ae15682
Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>